+++ source/coreload/stub/include/error.h
@@ -1,8 +1,10 @@
 #ifndef __ERROR_H__
 #define __ERROR_H__
 
-const char* error_string_15h(int);
-void print_error_15h(int);
+#include "types.h"
+
+const char* error_string_15h(int16_t);
+void print_error_15h(int16_t);
 
 #endif
